Refactor get_docstring to taskutils module
This will make it available to the lsst-task-topic directive code, and
other directives, as well.
