+++ documenteer/sphinxext/lssttasks/pyapisummary.py
@@ -8,13 +8,12 @@ from docutils import nodes
 from docutils.parsers.rst import Directive
 from sphinx.errors import SphinxError
 from sphinx.util.logging import getLogger
-from sphinx.util.inspect import getdoc, Signature
-from sphinx.util.docstrings import prepare_docstring
+from sphinx.util.inspect import Signature
 from sphinx.addnodes import (desc, desc_signature, desc_content, desc_addname,
                              desc_annotation, seealso)
 from sphinx.domains.python import _pseudo_parse_arglist, PyXRefRole
 
-from .taskutils import get_type
+from .taskutils import get_type, get_docstring, extract_docstring_summary
 
 
 class TaskApiDirective(Directive):
@@ -302,46 +301,3 @@ class TaskApiDirective(Directive):
         seealso_node += p_node
 
         return [seealso_node]
-
-
-def get_docstring(obj):
-    """Extract the docstring from an object as individual lines.
-
-    Parameters
-    ----------
-    obj : object
-        The Python object (class, function or method) to extract docstrings
-        from.
-
-    Returns
-    -------
-    lines : `list` of `str`
-        Individual docstring lines with common indentation removed, and
-        newline characters stripped.
-    """
-    docstring = getdoc(obj, allow_inherited=True)
-    # ignore is simply the number of initial lines to ignore when determining
-    # the docstring's baseline indent level. We really want "1" here.
-    return prepare_docstring(docstring, ignore=1)
-
-
-def extract_docstring_summary(docstring):
-    """Get the first summary sentence from a docstring.
-
-    Parameters
-    ----------
-    docstring : `list` of `str`
-        Output from `get_docstring`.
-
-    Returns
-    -------
-    summary : `str`
-        The plain-text summary sentence from teh docstring.
-    """
-    summary_lines = []
-    for line in docstring:
-        if line == '':
-            break
-        else:
-            summary_lines.append(line)
-    return ' '.join(summary_lines)
