+++ documenteer/sphinxext/lssttasks/taskutils.py
@@ -2,12 +2,15 @@
 """
 
 __all__ = ('get_task_config_class', 'get_task_config_fields',
-           'get_subtask_fields', 'typestring', 'get_type')
+           'get_subtask_fields', 'typestring', 'get_type',
+           'get_docstring', 'extract_docstring_summary')
 
 from importlib import import_module
 import inspect
 
 from sphinx.errors import SphinxError
+from sphinx.util.inspect import getdoc
+from sphinx.util.docstrings import prepare_docstring
 
 
 def get_task_config_class(task_name):
@@ -157,3 +160,46 @@ def typestring(obj):
     """
     obj_type = type(obj)
     return '.'.join((obj_type.__module__, obj_type.__name__))
+
+
+def get_docstring(obj):
+    """Extract the docstring from an object as individual lines.
+
+    Parameters
+    ----------
+    obj : object
+        The Python object (class, function or method) to extract docstrings
+        from.
+
+    Returns
+    -------
+    lines : `list` of `str`
+        Individual docstring lines with common indentation removed, and
+        newline characters stripped.
+    """
+    docstring = getdoc(obj, allow_inherited=True)
+    # ignore is simply the number of initial lines to ignore when determining
+    # the docstring's baseline indent level. We really want "1" here.
+    return prepare_docstring(docstring, ignore=1)
+
+
+def extract_docstring_summary(docstring):
+    """Get the first summary sentence from a docstring.
+
+    Parameters
+    ----------
+    docstring : `list` of `str`
+        Output from `get_docstring`.
+
+    Returns
+    -------
+    summary : `str`
+        The plain-text summary sentence from the docstring.
+    """
+    summary_lines = []
+    for line in docstring:
+        if line == '':
+            break
+        else:
+            summary_lines.append(line)
+    return ' '.join(summary_lines)
